+++ browser/src/lib.rs
@@ -1,5 +1,6 @@
 mod browser;
 mod engine;
+mod lex;
 mod ui;
 
 pub use browser::Browser;
